+++ src/test/java/net/kaczmarzyk/spring/data/jpa/web/SpecificationArgumentResolverTest.java
@@ -20,7 +20,6 @@ import net.kaczmarzyk.spring.data.jpa.domain.Like;
 import net.kaczmarzyk.spring.data.jpa.utils.QueryContext;
 import net.kaczmarzyk.spring.data.jpa.web.annotation.Join;
 import net.kaczmarzyk.spring.data.jpa.web.annotation.JoinFetch;
-import net.kaczmarzyk.spring.data.jpa.web.annotation.Joins;
 import net.kaczmarzyk.spring.data.jpa.web.annotation.Spec;
 import org.junit.Test;
 import org.springframework.core.MethodParameter;
@@ -167,30 +166,6 @@ public class SpecificationArgumentResolverTest extends ResolverTestBase {
                         new net.kaczmarzyk.spring.data.jpa.domain.JoinFetch<>(queryCtx, new String[]{ "fetch2" }, JoinType.LEFT, true)));
     }
     
-    @Test
-    public void resolvesJoinContainerForAnnotatedInterface() throws Exception {
-        MethodParameter param = MethodParameter.forExecutable(testMethod("testMethodWithCustomSpec_joinContainer", CustomSpecJoinContainer.class), 0);
-        NativeWebRequest req = mock(NativeWebRequest.class);
-        QueryContext queryCtx = new DefaultQueryContext();
-
-        when(req.getParameterValues("path1")).thenReturn(new String[] { "value1" });
-
-        Specification<?> resolved = (Specification<?>) resolver.resolveArgument(param, null, req, null);
-
-        assertThat(resolved)
-            .isInstanceOf(CustomSpecJoinContainer.class);
-
-        assertThat(innerSpecs(resolved))
-                .hasSize(2)
-                .contains(
-                        new Conjunction<>(
-                                new net.kaczmarzyk.spring.data.jpa.domain.JoinFetch<>(queryCtx, new String[]{"fetch1"}, JoinType.LEFT, true),
-                                new net.kaczmarzyk.spring.data.jpa.domain.JoinFetch<>(queryCtx, new String[]{"fetch2"}, JoinType.INNER, true)
-                        ),
-                        new Like<>(queryCtx, "path1", "value1")
-                );
-    }
-
     @Override
     protected Class<?> controllerClass() {
         return TestController.class;
@@ -207,14 +182,6 @@ public class SpecificationArgumentResolverTest extends ResolverTestBase {
     public static interface CustomSpecRepeatedFetch extends Specification<Object> {
     }
     
-    @Joins(fetch = {
-    	@JoinFetch(paths = { "fetch1" }),
-        @JoinFetch(paths = { "fetch2" }, joinType = JoinType.INNER)
-    })
-    @Spec(path = "path1", spec = Like.class)
-    public static interface CustomSpecJoinContainer extends Specification<Object> {
-    }
-    
     public static class TestController {
         
     	public void testMethodWithCustomSpec(CustomSpec spec) {
@@ -223,9 +190,6 @@ public class SpecificationArgumentResolverTest extends ResolverTestBase {
     	public void testMethodWithCustomSpec_repeatedFetch(CustomSpecRepeatedFetch spec) {
         }
     	
-    	public void testMethodWithCustomSpec_joinContainer(CustomSpecJoinContainer spec) {
-        }
-        
         public void testMethod(
                 @JoinFetch(paths = { "fetch1", "fetch2" })
                 @Spec(path = "path1", spec = Like.class) Specification<Object> spec) {
@@ -237,30 +201,23 @@ public class SpecificationArgumentResolverTest extends ResolverTestBase {
                 @Spec(path = "path1", spec = Like.class) Specification<Object> spec) {
         }
         
-        public void testMethod_joinContainerWithJoinFetch(
-                @Joins(fetch = {
-                	@JoinFetch(paths = { "fetch1" }),
-                    @JoinFetch(paths = { "fetch2" }, joinType = JoinType.INNER)
-                })
+        public void testMethod_repeatedJoinFetch(
+        		@JoinFetch(paths = { "fetch1" })
+                @JoinFetch(paths = { "fetch2" }, joinType = JoinType.INNER)
                 @Spec(path = "path1", spec = Like.class) Specification<Object> spec) {
         }
         
-        public void testMethod_joinContainerWithRegularJoin(
-                @Joins({
-                	@Join(path = "join1", alias = "alias1", type = JoinType.INNER, distinct = true),
-                	@Join(path = "join2", alias = "alias2", type = JoinType.LEFT, distinct = false)
-                })
+        public void testMethod_repeatedRegularJoin(
+        		@Join(path = "join1", alias = "alias1", type = JoinType.INNER, distinct = true)
+            	@Join(path = "join2", alias = "alias2", type = JoinType.LEFT, distinct = false)
                 @Spec(path = "path1", spec = Like.class) Specification<Object> spec) {
         }
         
-        public void testMethod_joinContainerWithRegularAndFetchJoins(
-                @Joins(value = {
-                	@Join(path = "join1", alias = "alias1", type = JoinType.INNER, distinct = true),
-                	@Join(path = "join2", alias = "alias2", type = JoinType.LEFT, distinct = false)
-                }, fetch = {
-                    	@JoinFetch(paths = { "fetch1" }),
-                        @JoinFetch(paths = { "fetch2" }, joinType = JoinType.INNER)
-                    })
+        public void testMethod_regularAndFetchJoins(
+        		@Join(path = "join1", alias = "alias1", type = JoinType.INNER, distinct = true)
+            	@Join(path = "join2", alias = "alias2", type = JoinType.LEFT, distinct = false)
+        		@JoinFetch(paths = { "fetch1" })
+                @JoinFetch(paths = { "fetch2" }, joinType = JoinType.INNER)
                 @Spec(path = "path1", spec = Like.class) Specification<Object> spec) {
         }
     }