+++ CHANGELOG.md
@@ -4,6 +4,11 @@ v3.0.0
   * Spring boot 3.0 is based on Hibernate version 6.X because in this version of hibernate all query results are distinct by default. This shouldn't affect most projects, but please be extra careful if you've ever used a spec with the `distinct=false` attribute.
 * Added support for spring native-image.
   * Specification-arg-resolver can be used in GraalVM native images, but it requires several additional configuration steps. This is due to the fact that this library relies on Java reflection heavily. Please see [README_native_image.md](README_native_image.md) for the details
+* Removed all deprecated items, including:
+  * DateAfter, DateAfterInclusive, DateBefore, DateBeforeInclusive - use GreaterThan, GreaterThanOrEqual, LessThan or LessThanOrEqual
+  * DateBetween - use Between
+  * EqualEnum - use Equal
+  * @Joins container annotation - use repeated @Join and @JoinFetch
 
 v2.14.0
 =======